+++ cpp/Stage.cpp
