+++ cpp/TextureManager.cpp
