+++ hpp/GameConfig.hpp
