+++ hpp/Stage.hpp
